requirements.txt
pyyaml
matplotlib==3.10.7
nltk==3.9.1
numpy==2.3.4
pandas==2.3.3
scikit_learn==1.7.2
tensorflow==2.15.1
tensorflow_intel==2.15.1
torch==2.8.0
